refactor result tests and change init-job
